Addition Refactorings
* Renamed "java-app" to "java-exec" so that it matches the java:exec goal of the maven-java-plugin
* Cleaned up generator names to not contain "."
