+++ core/pom.xml
@@ -52,11 +52,6 @@
       <artifactId>docker-maven-plugin</artifactId>
     </dependency>
 
-    <dependency>
-      <groupId>org.eclipse.jgit</groupId>
-      <artifactId>org.eclipse.jgit</artifactId>
-    </dependency>
-
   </dependencies>
 
   <build>
