+++ enricher/api/pom.xml
@@ -50,5 +50,10 @@
       <artifactId>fabric8-maven-core</artifactId>
     </dependency>
 
+    <dependency>
+      <groupId>org.eclipse.jgit</groupId>
+      <artifactId>org.eclipse.jgit</artifactId>
+    </dependency>
+
   </dependencies>
 </project>
