+++ enricher/cd/src/main/resources/META-INF/fabric8/enricher-default
