--- generator/java-app/src/main/resources/META-INF/fabric8/generator-default
@@ -1,1 +1,0 @@
-io.fabric8.maven.generator.javaapp.JavaAppGenerator
