+++ generator/java-exec/src/main/resources/META-INF/fabric8/generator-default
@@ -1,0 +1,1 @@
+io.fabric8.maven.generator.javaapp.JavaExecGenerator
