+++ generator/java-exec/src/main/resources/assemblies/java-app.xml
