+++ plugin/pom.xml
@@ -89,7 +89,7 @@
 
     <dependency>
       <groupId>io.fabric8</groupId>
-      <artifactId>fabric8-maven-generator-java-app</artifactId>
+      <artifactId>fabric8-maven-generator-java-exec</artifactId>
     </dependency>
 
     <dependency>
